requirements.txt
amqp==5.0.6
aiohttp==3.8.1
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==4.0.2
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
cfgv==3.3.1
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
distlib==0.3.2
Django==3.2.7
django-environ==0.8.1
filelock==3.0.12
geojson==2.5.0
gunicorn==20.1.0
identify==2.2.15
idna==3.2
kombu==5.1.0
lxml==4.6.3
nodeenv==1.6.0
packaging==21.3
platformdirs==2.3.0
pre-commit==2.15.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pyowm==3.2.0
pyparsing==3.0.8
PySocks==1.7.1
python-telegram-bot==13.11
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.7
vine==5.0.0
virtualenv==20.8.0
wcwidth==0.2.5
wrapt==1.14.0
